Refactor dragHorizon function and remove unused mouseEvent and time utilities
