+++ client/app/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './customFetch'
-export * from './mouseEvent'
-export * from './time'
+export * from './common'
